fix: redefine the question, point calculation, text style and some stuff
Closes #77
