+++ backend/src/controllers/userController.js
@@ -1,0 +1,10 @@
+import db from '../models/index.js';
+
+export const getUsers = async (req, res) => {
+    try {
+        const users = await db.User.findAll();
+        res.json(users);
+    } catch (error) {
+        res.status(500).json({ message: 'Failed to fetch users' });
+    }
+};
