+++ frontend/src/pages/Login.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import NavBar from '../components/NavBar';
+import Navbar from '../components/Navbar/NavBar';
 
 export default function Login() {
     return (
         <>
             <div className="w-full">
-                <NavBar />
+                <Navbar />
             </div>
             <div className="flex justify-center items-center h-screen bg-gray-100">
                 <div className="bg-white p-8 rounded-lg shadow-md w-full max-w-md">
